feat: US-005 - Move guards, decorators, and strategies to libs
