+++ src/app.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Get } from '@nestjs/common';
 import { AppService } from './app.service';
-import { Public } from './auth/decorators/public.decorator';
+import { Public } from '../libs/auth/decorators/public.decorator';
 
 @Controller()
 export class AppController {
