+++ src/auth/auth.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { UsersModule } from '../users/users.module';
 import { AuthModule as LibsAuthModule } from '../../libs/auth/auth.module';
 import { AuthController } from './auth.controller';
-import { JwtStrategy } from './strategies/jwt.strategy';
+import { JwtStrategy } from '../../libs/auth/strategies/jwt.strategy';
 import { UsersService } from '../users/users.service';
 import type { UserLookupProvider } from '../../libs/auth/interfaces/user-lookup-provider.interface';
